requirements.txt
yfinance
python-telegram-bot
python-dotenv
tabulate 